Use a semver library to parse versions
fly login would die with an error:
error: strconv.ParseInt: parsing "0+foo.1": invalid syntax
This is because it split the version on three dot-separated components,
and assumes they are all numbers.

Replace this inflexible parsing with a library that was built to parse
versions with pre- and post-release version segment.

We still assume the main release segment has 3 integer components, since
this seems in keeping with Concourse's versioning practice.

Signed-off-by: David Sharp <dsharp@pivotal.io>