Refactored `hijack` to `intercept` for brevity.
The principle of the "hijack" command was to intercept a running
container and modify it's runtime. While hijacking can be accurate from
a point of view, intercepting is the intended purpose and specifically
describes the action taken. Hijacking, especially in the context of
aerospace, is a socially murky topic due to numerous events in the past
20 years and can be negatively received by consumers of the tool.

Signed-off-by: Mike Lloyd <kevin.michael.lloyd@gmail.com>